refactor(shared-signals): one holder names both validation profile keys
SsfReceiverValidation named one profile and could name only one, so a second consumer of the
security-event core had nowhere to put its key. The holder becomes SharedSignalsValidationProfiles
and carries both: SecurityEvent, which the receiver registers and resolves, and LogoutToken, the
key a back-channel logout profile is composed under.

Nothing registers a logout profile today. The client's own receiver validates a logout token
directly on the JWT layer rather than through a profile, and the profile in the security-event
suite is a demonstration that the composite reshapes without the package changing - so the key
names a profile a host composes, not one this package ships.

That suite reaches the key across a package boundary, which is why it now references Shared
Signals; the reference exists for the constants and nothing else, and its item group says so.

Two XML crefs lost their target in the rename and pointed at a name no assembly declares, which
fails the build rather than degrading quietly - both now name the constant the receiver uses.
